Add support for Heroku
Heroku is a popular PaaS. This commit add a Procfile and a govendor
config file.

The Procfile is used to describe the workers that will be run on Heroku
platform. Here a web worker which launch ipd listening on $PORT (this
variable will be set by Heroku) and Trust the X-Forwarded-For header.

The vendor file is needed to use the Heroku go-build-pack. The build
pack is in charge to convert the source code to an executable program.